Revert "Refactored code to use user controllers and route"
This reverts commit 61d33087583719532cf303fceebffec8ef97c0fd.
