+++ model/userModel.js
@@ -1,5 +1,4 @@
 const mongoose = require('mongoose');
-const bcrypt = require('bcrypt');
 
 const Schema = mongoose.Schema;
 const ObjectId = Schema.ObjectId;
@@ -25,24 +24,6 @@ const UserSchema = new Schema({
         default: "user"
     }
 })
-
-UserSchema.pre(
-    'save',
-    async function (next){
-        const user = this;
-        const hash = await bcrypt.hash(this.password, 10);
-
-        this.password = hash;
-        next();
-    }
-)
-
-UserSchema.methods.isValidPassword = async function(password){
-    const user = this;
-    const compare = await bcrypt.compare(password, user.password);
-
-    return compare;
-}
 
 const userModel = mongoose.model('User', UserSchema);
